refactor: move ai utils to own folder and update imports
