+++ components/editor/ai/chat-interface.tsx
@@ -3,13 +3,13 @@
 import { toast } from "@/components/ui/use-toast";
 import { useAIThemeGeneration } from "@/hooks/use-ai-theme-generation";
 import { usePostLoginAction } from "@/hooks/use-post-login-action";
-import { buildPrompt } from "@/lib/ai-theme-generator";
+import { buildPrompt } from "@/lib/ai/ai-theme-generator";
 import { authClient } from "@/lib/auth-client";
 import { cn } from "@/lib/utils";
 import { useAIChatStore } from "@/store/ai-chat-store";
 import { useAuthStore } from "@/store/auth-store";
 import { AIPromptData } from "@/types/ai";
-import { attachLastGeneratedThemeMention, mentionsCount } from "@/utils/ai-prompt";
+import { attachLastGeneratedThemeMention, mentionsCount } from "@/utils/ai/ai-prompt";
 import dynamic from "next/dynamic";
 import { ChatInput } from "./chat-input";
 import { ClosableSuggestedPillActions } from "./closeable-suggested-pill-actions";
