+++ components/editor/ai/no-messages-placeholder.tsx
@@ -7,8 +7,8 @@ import { authClient } from "@/lib/auth-client";
 import { cn } from "@/lib/utils";
 import { useThemePresetStore } from "@/store/theme-preset-store";
 import { AIPromptData } from "@/types/ai";
-import { createCurrentThemePrompt } from "@/utils/ai-prompt";
-import { PROMPTS } from "@/utils/prompts";
+import { createCurrentThemePrompt } from "@/utils/ai/ai-prompt";
+import { PROMPTS } from "@/utils/ai/prompts";
 import { Blend, PaintRoller, WandSparkles } from "lucide-react";
 import { ComponentProps, Fragment } from "react";
 import TabsTriggerPill from "../theme-preview/tabs-trigger-pill";
