+++ utils/ai/ai-prompt.tsx
